refact: Refactoring all scaffold, modify role user
