+++ prisma/schema.prisma
@@ -84,20 +84,21 @@ model CartItem {
 }
 
 model User {
-  id       Int     @id @default(autoincrement())
-  name     String? @default("user name")
-  lastname String? @default("user lastname")
-  email    String  @unique
-  address  String
-  password String
-  role     Role[]  @default([USER])
-  cart     Cart[]
+  id        Int      @id @default(autoincrement())
+  name      String?  @default("user name")
+  lastname  String?  @default("user lastname")
+  email     String   @unique
+  phone     Int      @unique
+  address   String
+  password  String
+  createdAt DateTime @default(now())
+  updatedAt DateTime @default(now())
+  role      Role[]   @default([USER])
+  cart      Cart[]
 }
 
 enum Role {
   ADMIN
-  MANAGER
-  GUEST
   USER
 }
