+++ src/app.module.ts
@@ -1,13 +1,13 @@
 import { Module } from '@nestjs/common';
-import { ProductsModule } from './modules/products/products.module';
-import { AuthModule } from './modules/auth/auth.module';
-import { CategoriesModule } from './modules/categories/categories.module';
-import { PaymentsModule } from './modules/payments/payments.module';
+import { ProductsModule } from './features/products/products.module';
+import { AuthModule } from './features/auth/auth.module';
+import { CategoriesModule } from './features/categories/categories.module';
+import { PaymentsModule } from './features/payments/payments.module';
 import { OrdersModule } from './integrations/paypal/orders/orders.module';
-import { CartModule } from './modules/cart/cart.module';
-import { UsersModule } from './modules/users/users.module';
+import { CartModule } from './features/cart/cart.module';
+import { UsersModule } from './features/users/users.module';
 import { ThrottlerModule } from '@nestjs/throttler';
-import { SubcategoryModule } from './modules/subcategory/subcategory.module';
+import { SubcategoryModule } from './features/subcategory/subcategory.module';
 import { GraphQLModule } from '@nestjs/graphql';
 import { ApolloDriver, ApolloDriverConfig } from '@nestjs/apollo';
 import { join } from 'path';
