+++ src/features/auth/auth.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common';
 import { AuthService } from './auth.service';
 import { AuthController } from './auth.controller';
-import { UsersModule } from 'src/modules/users/users.module';
+import { UsersModule } from 'src/features/users/users.module';
 import { JwtModule } from '@nestjs/jwt';
 
 @Module({
