+++ src/features/auth/common/enums/role.enum.ts
@@ -1,6 +1,4 @@
 export enum Role {
   USER = 'USER',
   ADMIN = 'ADMIN',
-  MANAGER = 'MANAGER',
-  GUEST = 'GUEST',
 }
