+++ src/features/auth/dto/create-auth.dto.ts
@@ -1,0 +1,3 @@
+import { CreateUserDto } from 'src/features/users/dto/create-user.dto';
+
+export class CreateAuthDto extends CreateUserDto {}
