+++ src/features/categories/entity/category.entity.ts
@@ -1,6 +1,6 @@
 import { Field, Int, ObjectType } from '@nestjs/graphql';
-import { Product } from 'src/modules/products/entity/product.entity';
-import { Subcategory } from 'src/modules/subcategory/entities/subcategory.entity';
+import { Product } from 'src/features/products/entity/product.entity';
+import { Subcategory } from 'src/features/subcategory/entities/subcategory.entity';
 
 @ObjectType()
 export class Category {
