+++ src/features/products/entity/product.entity.ts
@@ -1,6 +1,5 @@
 import { Field, Float, Int, ObjectType } from '@nestjs/graphql';
-import { Cart } from 'src/modules/cart/entity/cart.entity';
-import { Category } from 'src/modules/categories/entity/category.entity';
+import { Category } from 'src/features/categories/entity/category.entity';
 
 @ObjectType()
 export class Product {
