+++ src/features/subcategory/entities/subcategory.entity.ts
